input: use joystick for moving, pointer for rotation
Implement a control style a la "The Conduit", where the pointer is used
to rotate the camera, and the joystick is used for moving. This still
needs some work (we probably want to show a small pointer on the Wii
screen, because currently it's hard to know where exactly you are
pointing at), but should give an idea. Having playing "The COnduit" for
a while, this feels much more natural to me.

This depends on #17 in order to be tested on the PC.

Addresses issue #1.